requirements.txt
faiss-cpu
sentence-transformers
llama-cpp-python
pypdf
numpy
pandas